requirements.txt
python-telegram-bot==12.2.0
requests==2.18.1